+++ src/main/java/at/ac/tuwien/sepm/assignment/individual/rest/dto/HorseJockeyCombinationDto.java
@@ -2,24 +2,23 @@ package at.ac.tuwien.sepm.assignment.individual.rest.dto;
 
 import java.time.LocalDateTime;
 import java.util.Objects;
-import java.math.BigDecimal;
 
 public class HorseJockeyCombinationDto {
     private Integer id;
     private Integer rank;
     private String horseName;
     private String jockeyName;
-    private BigDecimal avgSpeed;
-    private BigDecimal horseSpeed;
-    private BigDecimal skill;
-    private BigDecimal luckFactor;
+    private Double avgSpeed;
+    private Double horseSpeed;
+    private Double skill;
+    private Double luckFactor;
 
     public HorseJockeyCombinationDto() {
     }
 
     public HorseJockeyCombinationDto(Integer id, Integer rank, String horseName, 
-            String jockeyName, BigDecimal avgSpeed, BigDecimal horseSpeed,
-            BigDecimal skill, BigDecimal luckFactor) {
+            String jockeyName, Double avgSpeed, Double horseSpeed,
+            Double skill, Double luckFactor) {
         this.id = id;
         this.rank = rank;
         this.horseName = horseName;
@@ -62,35 +61,35 @@ public class HorseJockeyCombinationDto {
         this.jockeyName = jockeyName;
     }
 
-    public BigDecimal getAvgSpeed() {
+    public Double getAvgSpeed() {
         return avgSpeed;
     }
 
-    public void setAvgSpeed(BigDecimal avgSpeed) {
+    public void setAvgSpeed(Double avgSpeed) {
         this.avgSpeed = avgSpeed;
     }
 
-    public BigDecimal getHorseSpeed() {
+    public Double getHorseSpeed() {
         return horseSpeed;
     }
 
-    public void setHorseSpeed(BigDecimal horseSpeed) {
+    public void setHorseSpeed(Double horseSpeed) {
         this.horseSpeed = horseSpeed;
     }
 
-    public BigDecimal getSkill() {
+    public Double getSkill() {
         return skill;
     }
 
-    public void setSkill(BigDecimal skill) {
+    public void setSkill(Double skill) {
         this.skill = skill;
     }
 
-    public BigDecimal getLuckFactor() {
+    public Double getLuckFactor() {
         return luckFactor;
     }
 
-    public void setLuckFactor(BigDecimal luckFactor) {
+    public void setLuckFactor(Double luckFactor) {
         this.luckFactor = luckFactor;
     }
